requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # For YOLO
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0  # For NLP tasks
spacy>=3.6.0
nltk>=3.8.1
pillow>=10.0.0
tqdm>=4.65.0
librosa>=0.10.0  # For audio analysis
soundfile>=0.12.1  # For audio file handling
scipy>=1.10.0  # Required by librosa
streamlit>=1.32.0  # For the web interface
plotly>=5.18.0  # For interactive visualizations
watchdog>=3.0.0  # For better file watching performance 